+++ spec/project_spec.rb
@@ -165,6 +165,41 @@ module ProjectSpecs
         new_instance.should == @project
       end
 
+      it "can save a project after removing a subproject" do
+        project = Xcodeproj::Project.open(fixture_path("Sample Project/Cocoa Application.xcodeproj"))
+
+        # UUID's related to ReferencedProject.xcodeproj (subproject)
+        # FIXME: these UUID's should all be deleted automatically when calling 
+        #        remove_from_project on the subproject PBXFileReference
+        #        See https://github.com/CocoaPods/Xcodeproj/issues/158
+        uuids_to_remove = [
+          "E5FBB3451635ED35009E96B0", # The Xcode subproject file reference that should trigger the removal.
+
+          "5138059B16499F4C001D82AD", # PBXContainerItemProxy links to E5FBB3451635ED35009E96B0
+          "5138059C16499F4C001D82AD", # PBXTargetDependency links to 5138059B16499F4C001D82AD
+          "E5FBB3461635ED35009E96B0", # PBXGroup for products links to E5FBB34C1635ED36009E96B0, E5FBB34E1635ED36009E96B0, E5FBB3501635ED36009E96B0
+
+          "E5FBB34B1635ED36009E96B0", # PBXContainerItemProxy links to E5FBB3451635ED35009E96B0
+          "E5FBB34C1635ED36009E96B0", # PBXReferenceProxy links to E5FBB34B1635ED36009E96B0
+
+          "E5FBB34D1635ED36009E96B0", # PBXContainerItemProxy links to E5FBB3451635ED35009E96B0
+          "E5FBB34E1635ED36009E96B0", # PBXReferenceProxy links to E5FBB34D1635ED36009E96B0
+
+          "E5FBB34F1635ED36009E96B0", # PBXContainerItemProxy links to E5FBB3451635ED35009E96B0
+          "E5FBB3501635ED36009E96B0", # PBXReferenceProxy links to E5FBB34F1635ED36009E96B0
+        ]
+
+        subproject_file_reference = project.objects_by_uuid['E5FBB3451635ED35009E96B0']
+        subproject_file_reference.remove_from_project
+        project.save(@path)
+
+        new_instance = Xcodeproj::Project.open(@path)
+        new_instance.objects.count.should > 0 # make sure we still have a valid project
+        new_instance.root_object.project_references.should.be.empty # this contains the Products group of the external project
+        removed_objects = new_instance.objects.select { |o| uuids_to_remove.include?(o.uuid) }
+        removed_objects.count.should == 0
+      end
+
       it "can open a project and save it without altering any information" do
         project = Xcodeproj::Project.open(fixture_path("Sample Project/Cocoa Application.xcodeproj"))
         plist = Xcodeproj.read_plist(fixture_path("Sample Project/Cocoa Application.xcodeproj/project.pbxproj"))